Refactor: 변경된 구조에 회원가입 적용
